+++ dapps/dapp-box/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "conflux-hub",
+  "name": "dapp-box",
   "private": true,
   "version": "0.0.0",
   "scripts": {
