+++ dapps/dapp-box/src/modules/Sidebar/index.tsx
@@ -3,10 +3,10 @@ import { a, useSpring, config } from '@react-spring/web';
 import { useNavigate } from 'react-router-dom';
 import cx from 'clsx';
 import Popper from 'common/components/Popper';
-import useCurrentDapp from 'conflux-hub/src/hooks/useCurrentDapp';
-import Logo from 'conflux-hub/src/assets/logo.svg';
-import Expand from 'conflux-hub/src/assets/expand.svg';
-import { dapps } from 'conflux-hub/src/App';
+import useCurrentDapp from 'dapp-box/src/hooks/useCurrentDapp';
+import Logo from 'dapp-box/src/assets/logo.svg';
+import Expand from 'dapp-box/src/assets/expand.svg';
+import { dapps } from 'dapp-box/src/App';
 import './index.css';
 
 const Sidebar: React.FC = () => {
